create payment link & handle order status if payment or run out of time

+++ WebApplication1/Models/OrderItems.cs
@@ -8,7 +8,7 @@
         public MenuItems MenuItems { get; set; }
         public Guid MenuItemsId { get; set; }
         public int Quantity { get; set; }
-        public double UnitPrice { get; set; }
-        public double SubTotal { get; set; }
+        public int UnitPrice { get; set; }
+        public int SubTotal { get; set; }
     }
 }
